# openapi.yaml
openapi: 3.0.3
info:
  title: MyGoLangApp API
  version: "1.0.0"
servers:
  - url: http://localhost:8080
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    search:
      name: search
      in: query
      description: Free-text search term for list filtering
      required: false
      schema:
        type: string
        example: "toko"
    sort:
      name: sort
      in: query
      description: >
        Comma-separated sort fields. Common patterns:
        `-created_at` (prefix `-` = desc)
        `amount` (no prefix `-` = asc)
      required: false
      schema:
        type: string
        example: "-created_at"
    page:
      name: page
      in: query
      description: Page number (1-based)
      required: false
      schema:
        type: integer
        default: 1
        minimum: 1
    limit:
      name: limit
      in: query
      description: Number of items per page (max 100)
      required: false
      schema:
        type: integer
        default: 10
        minimum: 1
        maximum: 100
  schemas:
    Error:
      type: object
      properties:
        code:
          type: integer
          example: 401
        message:
          type: string
          example: "Unauthenticated: missing or invalid token"
      required:
        - code
        - message
...
    Payment:
      type: object
      properties:
        id:
          type: string
        merchant_id:
          type: integer
        merchant_name:
          type: string
        status:
          type: string
        amount:
          type: string
        created_at:
          type: string
          format: date-time
    CreatePaymentRequest:
      type: object
      required: [merchant_id, amount, status]
      properties:
        merchant_id:
          type: integer
        amount:
          type: string
        status:
          type: string
...
  responses:
...
    PaymentListResponse:
      description: Payment List
      content:
        application/json:
          schema:
            type: object
            properties:
              payments:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
...
    UnauthorizedError:
      description: Authentication failed or missing credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            unauthenticated:
              summary: Missing or invalid token
              value:
                code: 401
                message: "missing auth header"
    BadRequestError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
paths:
...
  /dashboard/v1/payments:
    get:
      tags: [Payment]
      summary: List payments
      operationId: getDashboardV1Payments
      parameters:
        - $ref: "#/components/parameters/search"
        - $ref: "#/components/parameters/sort"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/limit"
        - in: query
          name: status
          schema:
            type: string
          description: status of payment (completed, processing, or failed)
        - in: query
          name: id
          schema:
            type: string
          description: payment id
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/PaymentListResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    post:
      tags: [Payment]
      summary: Create a payment
      operationId: postDashboardV1Payments
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentRequest"
      responses:
        "201":
          description: Created payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...